I had added the README file to descibe a bit more my code on runnerup.py
For runnerup.py I used sort to order the items from smallest to biggest.
After that I popped the last item and assume that it is the largest
Then I pop the next element in the list. If it is not equal to the max
then that is the second/runner up and the loop ends.